Project 7: Convert weight fields from int to decimal (#2466)
- Change EventSummary.PickedWeight from int to decimal
- Update Stats.TotalWeightInPounds/Kilograms to decimal
- Add database migration with decimal(10,1) precision
- Update EventSummaryManager weight calculations with proper conversion
- Add weight entry UI to mobile EditEventSummaryPage
- Add weight display column to mobile ViewEventSummaryPage
- Support user preference for default weight unit (lbs/kg)

+++ TrashMob.Models/EventSummary.cs
@@ -40,7 +40,7 @@ namespace TrashMob.Models
         /// <summary>
         /// Gets or sets the total weight of trash picked up during the event.
         /// </summary>
-        public int PickedWeight { get; set; }
+        public decimal PickedWeight { get; set; }
 
         /// <summary>
         /// Gets or sets the identifier of the weight unit used for the picked weight.
